+++ BXFW/Scripts/Editor/General/Utility/EditorGUIAdditionals.cs
@@ -1,7 +1,7 @@
 using System;
+using System.Collections.Generic;
 using UnityEngine;
 using UnityEditor;
-using System.Collections.Generic;
 
 namespace BXFW.Tools.Editor
 {
