+++ BXFW/Scripts/Main/General/Utility/Attributes/BigSpriteFieldAttribute.cs
@@ -1,0 +1,25 @@
+using System;
+using UnityEngine;
+
+namespace BXFW
+{
+    /// <summary>
+    /// Attribute to draw <see cref="Sprite"/> fields with a big preview texture.
+    /// </summary>
+    [AttributeUsage(AttributeTargets.Field, Inherited = true, AllowMultiple = false)]
+    public class BigSpriteFieldAttribute : PropertyAttribute
+    {
+#if UNITY_EDITOR
+        public readonly float spriteBoxRectHeight = 44f;
+#endif
+        public BigSpriteFieldAttribute()
+        { }
+
+        public BigSpriteFieldAttribute(float spriteHeight)
+        {
+#if UNITY_EDITOR
+            spriteBoxRectHeight = spriteHeight;
+#endif
+        }
+    }
+}
